requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
pandas==2.1.3
python-multipart==0.0.9
python-dotenv==1.0.0
openai-agents==0.2.11
openai==1.104.1
daytona>=0.26.0
